workflow.go
package app

import (
	"time"

	"go.temporal.io/sdk/temporal"
	"go.temporal.io/sdk/workflow"
)

func DownloadWorkflow(ctx workflow.Context, url string, dir string) (string, error) {
	retrypolicy := &temporal.RetryPolicy{
		InitialInterval:    time.Second,
		BackoffCoefficient: 2.0,
		MaximumInterval:    time.Second * 100,
	}

	options := workflow.ActivityOptions{
		StartToCloseTimeout: time.Second * 120,
		RetryPolicy:         retrypolicy,
	}

	ctx = workflow.WithActivityOptions(ctx, options)

	var result string
	err := workflow.ExecuteActivity(ctx, TryDownload, url, dir).Get(ctx, &result)

	return result, err
}
